web/b01lers_casino/src/requirements.txt
flask==3.0.2
pyjwt==2.8.0
pyopenssl
